% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{d3plusActive}
\alias{d3plusActive}
\title{Adds .active() method to the chart}
\usage{
d3plusActive(d3plus, value = NULL, ...)
}
\description{
Function to define the key associated with the active parts .
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#active}{d3plus wiki page}.
}
