% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggs.R
\name{d3plusAggs}
\alias{d3plusAggs}
\title{Adds .aggs() method to the chart}
\usage{
d3plusAggs(d3plus, ...)
}
\description{
Defines how specific variables should be aggregated.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#aggs}{d3plus wiki page}.
}

