% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrs.R
\name{d3plusAttrs}
\alias{d3plusAttrs}
\title{Adds .attrs() method to the chart}
\usage{
d3plusAttrs(d3plus, value = NULL, ...)
}
\description{
Pass data with attributes. Use a named list correspondent to id values.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#attrs}{d3plus wiki page}.
}
