% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{d3plusAxes}
\alias{d3plusAxes}
\title{Adds .axes() method to the chart}
\usage{
d3plusAxes(d3plus, ticks = TRUE, mirror = FALSE, ...)
}
\description{
Defines specifics aspects of the X and Y axes.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#axes}{d3plus wiki page}.
}

