% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plusbar.R
\name{d3plusbar}
\alias{d3plusbar}
\title{D3plus bar chart}
\usage{
d3plusbar(data, x, y, id = NULL, xlabel = FALSE, ylabel = FALSE,
  title = FALSE, xtime = FALSE, attributes = NULL, gridx = FALSE,
  gridy = TRUE, filters = TRUE, background_color = "#FFFFFF",
  width_panel = 0, legend = TRUE, stackedy = c(FALSE, FALSE),
  stackedx = c(FALSE, FALSE), scaley = "linear", scalex = "discrete",
  legend_data = FALSE, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data.frame object to create the visualization.}

\item{x}{name of the column for x-axis.}

\item{y}{name of the column for y-axis.}

\item{id}{a vector with the id's column names.}

\item{xlabel}{label of x-axis. Also it could be TRUE or FALSE. If true the value in `x`param is used. If false, x-label is ommited.}

\item{ylabel}{label of y-axis. Also it could be TRUE or FALSE. If true the value in `y`param is used. If false, y-label is ommited.}

\item{title}{title of the visualization. If false, no title is displayed.}

\item{xtime}{a logical value indicating whether or not the x variable is a time variable.}

\item{attributes}{a named list with attributes data.frame. If the visualization has more than one level, each level is should be an element in the list. See the examples.}

\item{gridx}{a logical value indicating whether or not the grid lines of x-axis should be displayed.}

\item{gridy}{a logical value indicating whether or not the grid lines of y-axis should be displayed.}

\item{filters}{a logical value indicanting whether or not the filter buttons should be included in the legend's tooltips.}

\item{background_color}{a hex color value for the visualization background.}

\item{width_panel}{a number indicating the panel's border width.}

\item{legend}{a logical value indicating whether or not the legend is visible.}
}
\description{
Function to create D3plus bar charts.
}
\examples{

}

