% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{d3plusCols}
\alias{d3plusCols}
\title{Adds .cols() method to the chart}
\usage{
d3plusCols(d3plus, value = NULL, ...)
}
\description{
Specify the columns in a table visualization.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#cols}{d3plus wiki page}.
}
