% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{d3plusCoords}
\alias{d3plusCoords}
\title{Adds .coords() method to the chart}
\usage{
d3plusCoords(d3plus, value = NULL, ...)
}
\description{
Function to pass topojson object to D3plus. It could be a loaded topojson file or a URL.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#coords}{d3plus wiki page}.
}
