% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{d3plusData}
\alias{d3plusData}
\title{Adds .data() method to the chart}
\usage{
d3plusData(d3plus, value = NULL, ...)
}
\description{
Sets the data to the visualization and defines other options. If the user does not use any extra option, the data can be informed as an argument in d3plus() function.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#data}{d3plus wiki page}.
}
