% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{d3plusDepth}
\alias{d3plusDepth}
\title{Adds .depth() method to the chart}
\usage{
d3plusDepth(d3plus, value = NULL, ...)
}
\description{
If `id` has more than one level, this function sets a specific nesting level.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#depth}{d3plus wiki page}.
}

