% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{d3plusEdges}
\alias{d3plusEdges}
\title{Adds .edges() method to the chart}
\usage{
d3plusEdges(d3plus, value = NULL, ...)
}
\description{
Sets the list of the edges in a network or rings charts. The data.frame should have the source and targe columns. The values should match with the id.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#edges}{d3plus wiki page}.
}
