% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus.R
\name{d3plusFocus}
\alias{d3plusFocus}
\title{Adds .focus() method to the chart}
\usage{
d3plusFocus(d3plus, value = NULL, ...)
}
\description{
If supported, it defines which data node to focus on.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#focus}{d3plus wiki page}.
}
