% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{d3plusFont}
\alias{d3plusFont}
\title{Adds .font() method to the chart}
\usage{
d3plusFont(d3plus, value = NULL, ...)
}
\description{
Sets global fonts options to the chart.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#font}{d3plus wiki page}.
}
