% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{d3plusFooter}
\alias{d3plusFooter}
\title{Adds .footer() method to the chart}
\usage{
d3plusFooter(d3plus, value = NULL, ...)
}
\description{
Defines the footer of the visualization.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#footer}{d3plus wiki page}.
}
