% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{d3plusFormat}
\alias{d3plusFormat}
\title{Adds .format() method to the chart}
\usage{
d3plusFormat(d3plus, value = NULL, ...)
}
\description{
Format numbers and texts displayed in the visualization.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#format}{d3plus wiki page}.
}
