% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{d3plusId}
\alias{d3plusId}
\title{Set other .id() options to the chart. The d3plus() function already sets the value to the .id() method.}
\usage{
d3plusId(d3plus, value = NULL, ...)
}
\description{
Defines the options to the data point's identifier.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#id}{d3plus wiki page}.
}
