% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{d3plusLabels}
\alias{d3plusLabels}
\title{Adds .labels() method to the chart}
\usage{
d3plusLabels(d3plus, value = NULL, ...)
}
\description{
Controls the labels settings. To control which text should be showed, use the d3plusText() function.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#labels}{d3plus wiki page}.
}
