% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{d3plusLegend}
\alias{d3plusLegend}
\title{Adds .legend() method to the chart}
\usage{
d3plusLegend(d3plus, value = TRUE, ...)
}
\description{
Adds and controls legend options.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#x}{d3plus wiki page}.
}

