% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{d3plusNodes}
\alias{d3plusNodes}
\title{Adds .nodes() method to the chart}
\usage{
d3plusNodes(d3plus, value = NULL, ...)
}
\description{
Sets the positions of the nodes (x, y) in a network.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#nodes}{d3plus wiki page}.
}

