% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{d3plusSize}
\alias{d3plusSize}
\title{Adds .size() method to the chart}
\usage{
d3plusSize(d3plus, value = NULL, ...)
}
\description{
Defines the size variable and controls other options.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#size}{d3plus wiki page}.
}
