% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{d3plusText}
\alias{d3plusText}
\title{Adds .text() method to the chart}
\usage{
d3plusText(d3plus, value = NULL, ...)
}
\description{
Defines the text to be shown in shapes and tooltips.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#text}{d3plus wiki page}.
}
