% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{d3plusTime}
\alias{d3plusTime}
\title{Adds .time() method to the chart}
\usage{
d3plusTime(d3plus, value = NULL, ...)
}
\description{
Defines the time variable and controls other options. The time variable slices the chart in time periods
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#time}{d3plus wiki page}.
}
