% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{d3plusTitle}
\alias{d3plusTitle}
\title{Adds .title() method to the chart}
\usage{
d3plusTitle(d3plus, value = NULL, ...)
}
\description{
Defines the visualization's title and control other options.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#title}{d3plus wiki page}.
}
