% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{d3plusTotal}
\alias{d3plusTotal}
\title{Adds .total() method to the chart}
\usage{
d3plusTotal(d3plus, value = NULL, ...)
}
\description{
Defines the variable in data associated with the total number of parts of a data point.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#total}{d3plus wiki page}.
}

