% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{d3plusUi}
\alias{d3plusUi}
\title{Adds .ui() method to the chart}
\usage{
d3plusUi(d3plus, value = NULL, ...)
}
\description{
Adds UI elements to the chart that can change values for a specific method.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#axes}{d3plus wiki page}.
}
