% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{d3plusX}
\alias{d3plusX}
\title{Adds.x() method to the chart}
\usage{
d3plusX(d3plus, value = NULL, ...)
}
\description{
Adds and controls the x axis options in the visualization.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#x}{d3plus wiki page}.
}

