% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{d3plusZoom}
\alias{d3plusZoom}
\title{Adds .zoom() method to the chart}
\usage{
d3plusZoom(d3plus, value = NULL, ...)
}
\description{
Enables or Disables visualization zooming.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#zoom}{d3plus wiki page}.
}
