% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plus.R
\name{d3plus}
\alias{d3plus}
\title{Create D3plus charts in R}
\usage{
d3plus(
  data = NULL,
  type = c("tree_map", "bar", "line", "scatter", "geo_map", "stacked", "radar",
    "sankey", "rings", "network", "box", "bubbles"),
  id = NULL,
  currency = "US$",
  number_text = c("K", "M", "M", "B", "B"),
  currency_var = NULL,
  percent_var = NULL,
  noformat_number_var = NULL,
  d3plus_number_format = TRUE,
  locale = "en_US",
  dictionary = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  clean_previous = FALSE
)
}
\arguments{
\item{data}{a data.frame object to create the visualization.}

\item{type}{a name indicating the type of the chart.}

\item{id}{a vector with the id's column names.}

\item{currency}{a name indicating the currency symbol (US$, R$, etc.).}

\item{number_text}{a name to show after the number. Example: c("K", "M", "M", "B", "B").}

\item{currency_var}{a vector with variable names to apply the currency format.}

\item{percent_var}{a vector with variable names to apply the share format.}

\item{noformat_number_var}{a vector variables names to not apply any formatting.}

\item{d3plus_number_format}{a logical indicating if the d3plus.number.format should be applied to numbers.}

\item{locale}{a name indicating the language to be used}

\item{dictionary}{a named list. The element's name should be the original text and the value should be the new text.}

\item{width}{a number (in pixels) or percent text ("100\%") indicating the visualization's width.}

\item{height}{a number (in pixels) or percent text ("100\%") indicating the visualization's height.}

\item{elementId}{a name for the visualization's id.}

\item{clean_previous}{a logical indicating whether the visualization is restarted after input update (only for shiny).}
}
\description{
A function to initialize a D3plus chart in R
}
\examples{


}
