% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plusbar.R
\name{d3plusbar}
\alias{d3plusbar}
\title{D3plus bar chart}
\usage{
d3plusbar(data, x, y, id = NULL, xlabel = FALSE, ylabel = FALSE,
  title = FALSE, xtime = FALSE, attributes = NULL, gridx = FALSE,
  gridy = TRUE, filters = TRUE, background_color = "#FFFFFF",
  width_panel = 0, legend = TRUE, tooltip = NULL, stackedy = c(FALSE,
  FALSE), stackedx = c(FALSE, FALSE), scaley = "linear",
  scalex = "discrete", legend_data = FALSE, legend_size = 20,
  currency = "US$", number_text = c("K", "M", "M", "B", "B"),
  currency_var = NULL, percent_var = NULL, locale = "en_US",
  dictionary = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data.frame object to create the visualization.}

\item{x}{name of the column for x-axis.}

\item{y}{name of the column for y-axis.}

\item{id}{a vector with the id's column names.}

\item{xlabel}{label of x-axis. Also it could be TRUE or FALSE. If true the value in `x`param is used. If false, x-label is ommited.}

\item{ylabel}{label of y-axis. Also it could be TRUE or FALSE. If true the value in `y`param is used. If false, y-label is ommited.}

\item{title}{title of the visualization. If false, no title is displayed.}

\item{xtime}{a logical value indicating whether or not the x variable is a time variable.}

\item{attributes}{a named list with attributes data.frame. If the visualization has more than one level, each level is should be an element in the list. See the examples.}

\item{gridx}{a logical value indicating whether or not the grid lines of x-axis should be displayed.}

\item{gridy}{a logical value indicating whether or not the grid lines of y-axis should be displayed.}

\item{filters}{a logical value indicanting whether or not the filter buttons should be included in the legend's tooltips.}

\item{background_color}{an hex color value for the visualization background.}

\item{width_panel}{a number indicating the panel's border width.}

\item{legend}{a logical value indicating whether or not the legend is visible.}

\item{stackedy}{a vector with logical value indicating whether or not at that level bars should be stacked. Use only when y scale is linear.}

\item{stackedx}{a vector with logical value indicating whether or not at that level bars should be stacked. Use only when x scale is linear.}

\item{scaley}{a name indicating the scale type for the y axis (log, linear, discrete and share).}

\item{scalex}{a name indicating the scale type for the x axis (log, linear, discrete and share).}

\item{legend_data}{a logical value indicating if the legend's tooltips should show some data summaries.}

\item{legend_size}{a number indicating the size of legend box.}

\item{currency}{a name indicating the currency symbol (US$, R$, etc.).}

\item{number_text}{a name to show after the number. In portuguese: c("Mil", "Milh찾o", "Milh천es", "Bilh찾o", "Bilh천es").}

\item{currency_var}{a vector with variable names to apply the currency format.}

\item{percent_var}{a vector with variable names to apply the share format.}

\item{locale}{a name indicating the language to be used}

\item{dictionary}{a named list. The element's name should be the original text and the value should be the new text.}

\item{width}{a number (in pixels) or percent text ("100\%") indicating the visualization's width.}

\item{height}{a number (in pixels) or percent text ("100\%") indicating the visualization's height.}

\item{elementId}{a name for the visualization's ID.}
}
\description{
Function to create D3plus bar charts.
}
\examples{

library(D3plusR)
library(dplyr)
data("trade_bra_chn")
# Fake shares
trade_bra_chn <- trade_bra_chn \%>\% 
  mutate(share = sample(100, nrow(trade_bra_chn), replace = TRUE))
dictionary <- list(TradeValue = "Trade Value", Period = "Year",
                   share = "Share")
attributes <- list(Trade.Flow = data.frame(Trade.Flow = c("Export", "Import"),
                                           hex = c("#344969", "#992234")))
d3plusbar(data = trade_bra_chn,
          x = "Period",
          y = "TradeValue",
          id = "Trade.Flow",
          tooltip = c("Period", "TradeValue", "share"),
          xtime = TRUE,
          height = 400,
          width = "100\%",
          title = "Brazilian Exports and Imports to/from China",
          currency_var = "TradeValue",
          dictionary = dictionary,
          attributes = attributes,
          elementId = 'viz1')


}

