% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y.R
\name{d3plusY}
\alias{d3plusY}
\title{Adds .y() method to the chart}
\usage{
d3plusY(d3plus, value = NULL, ...)
}
\description{
Adds and controls the y axis options in the visualization.
See options: \href{https://github.com/alexandersimoes/d3plus/wiki/Visualizations#x}{d3plus wiki page}.
}

